finalized typescript refactor w/ updated webpack and tsconfig
