+++ src/app/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App';
+import App from './components/AppDir/App';
+
 
 ReactDOM.render(<App />, document.getElementById('root'));
